♻️ #39 refactor(modal): 모달 UI 로직과 내부 콘텐츠 분리 - BookDetailModal의 책임을 단순화하여, 모달 UI 렌더링에 집중하도록 리팩토링   액션 관련 UI는 별도 컴포넌트로 분리하여 역할 분담
